Changed async thread static field to instance field. This decreases execution time at the cost of increased memory.

Master:

|         Type | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |    True | 2,367.0 ns | 12.25 ns | 10.86 ns |         - |    688 B |

This PR:

|         Type | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |    True | 2,241.8 ns |  9.61 ns |  8.99 ns |         - |    712 B |